% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_sims.R
\name{do_sim_multiple}
\alias{do_sim_multiple}
\title{Simulation calibration experiment for model}
\usage{
do_sim_multiple(object, x0, m, sd1 = sd1, sd2 = sd2, sims = 1000)
}
\arguments{
\item{object}{lm representing idealized calibration curve}

\item{x0}{unknown values of $x$ over which $y'$ is to be generated}

\item{m}{length of independent $y'$ measurement to be genrated for each $x_0$}

\item{sd1}{standard deviation of measurement error during calibration stage}

\item{sd2}{standard deviation of measurement error during prediction stage}

\item{sims}{number of iterations}
}
\value{
A list containing the results of the simulation and simulation settings with the following components: \itemize{
\item{results}{A \code{data.frame} with the estimate and prediction interval for each x0}
\item{data}{A \code{data.frame} of the \emph{true} calibration curve contained within the provided model}

}
}
\description{
Takes provides model, y = f(x) and simulates calibration experiment. Provided model object represents idealize model E[y] ~ x  and must contain data.frame
with response and covariates for fitting. \code{sd1}  and \code{sd2} represents standard deviation of measurement errors during the calibration stage and prediction stage, respectively.
For each iteration, responses in the calibration stage are generated by y_i = f(x_i) + N(0,\code{sd1}) for each $i = 1,...,n. During the prediction stage, each provided \code{x0} generates an observed response
y_j' = f(x_0) + N(0,\code{sd2}) for j = 1,...,m. The calibration curve, f(x), is estimated via \code{\link{lm}} where \code{formula(object)} for observed data (x_1,y_1), ... , (x_n,y_n). For each x_0, the estiamted model, hat(f)(x), is used to
estimate $x_0$ based on its corresponding observed set (y_1', y_2, ...., y_m').
}
