% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invpred.R
\name{invpred}
\alias{invpred}
\alias{invpred.lm}
\title{Inverse prediction for linear models}
\usage{
invpred(object, y0, ...)

\method{invpred}{lm}(
  object,
  y0,
  interval = c("welch", "t", "bf", "wald"),
  level = 0.95,
  mean.response = FALSE,
  x0.name,
  newdata,
  data,
  lower,
  upper,
  extendInt = "no",
  tol = .Machine$double.eps^0.25,
  maxiter = 1000,
  k
)
}
\arguments{
\item{object}{An object that inherits from class \code{\link[stats]{lm}}}

\item{y0}{The value of the observed response(s).  \code{y0} should be on
the scale of the response variable (e.g., a number between 0 and 1 for
binomial families).}

\item{...}{Additional optional arguments. At present, no optional arguments
are used.}

\item{interval}{The type of interval required.}

\item{level}{A numeric scalar between 0 and 1 giving the confidence level
for the interval to be calculated.}

\item{mean.response}{Logical indicating whether confidence intervals should
correspond to an individual response (\code{FALSE}) or a mean response
(\code{TRUE}). For \code{\link[stats]{glm}} objects, this is always
\code{TRUE}.}

\item{x0.name}{For multiple linear regression, a character string giving the
name of the predictor variable of interest.}

\item{newdata}{For multiple linear regression, a \code{data.frame} giving the
values of interest for all other predictor variables (i.e., those other than
\code{x0.name}).}

\item{data}{An optional data frame. This is required if \code{object$data} is
\code{NULL}.}

\item{lower}{The lower endpoint of the interval to be searched.}

\item{upper}{The upper endpoint of the interval to be searched.}

\item{extendInt}{Character string specifying if the interval
\code{c(lower, upper)} should be extended or directly produce an error when
the inverse of the prediction function does not have differing signs at the
endpoints. The default, \code{"no"}, keeps the search interval and hence
produces an error. Can be abbreviated. See the documentation for the
\code{base} R function \code{uniroot} for details.}

\item{tol}{The desired accuracy passed on to \code{\link[stats]{uniroot}}.
Recommend a minimum of \code{1e-10}.}

\item{maxiter}{The maximum number of iterations passed on to \code{uniroot}.}

\item{nsim}{Positive integer specifying the number of bootstrap simulations;
the bootstrap B (or R).}

\item{boot.type}{Character string specifying the type of bootstrap to use
when \code{interval = "percentile"}. Options are \code{"parametric"} and
\code{"nonparametric"}.}

\item{seed}{Optional argument to \code{\link{set.seed}}.}

\item{progress}{Logical indicating whether to display a text-based progress
bar during the bootstrap simulation.}

\item{q1}{Optional lower cutoff to be used in forming confidence intervals.
Only used when \code{object} inherits from class \code{\link[nlme]{lme}}.
Defaults to \code{stats::qnorm((1+level)/2)}.}

\item{q2}{Optional upper cutoff to be used in forming confidence intervals.
Only used when \code{object} inherits from class \code{\link[nlme]{lme}}.
Defaults to \code{stats::qnorm((1-level)/2)}.}
}
\value{
Returns an object of class \code{"invest"} or, if
\code{interval = "percentile"}, of class \code{c("invest", "bootCal")}. The
generic function \code{{plot}} can be used to plot the output
of the bootstrap simulation when \code{interval = "percentile"}.

  An object of class \code{"invest"} containing the following components:
  \itemize{
    \item \code{estimate} The estimate of x0.
    \item \code{lwr} The lower confidence limit for x0.
    \item \code{upr} The upper confidence limit for x0.
    \item \code{se} An estimate of the standard error (Wald and percentile
                    intervals only).
    \item \code{bias} The bootstrap estimate of bias (percentile interval
                      only).
    \item \code{bootreps} Vector of bootstrap replicates (percentile
                          interval only).
    \item \code{nsim} The number of bootstrap replicates (percentile
                      interval only).
    \item \code{interval} The method used for calculating \code{lower} and
          \code{upper} (only used by \code{{print}} method).
  }
}
\description{
Provides point estimate and prediction intervals for unknown predictor value(s)
that corresponds to its observed proxy measurement or specified value of the mean response.
}
